Move collection to new file
